Outsource common setup_axis method
